requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
factory_boy==3.3.3
Faker==37.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
packaging==25.0
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylint==3.3.6
pytest==8.3.5
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tomlkit==0.13.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
